+++ app/services/activity/import.rb
@@ -23,6 +23,21 @@ class Activity
         headings
       end
 
+      def invalid_non_oda_attribute_errors(activity:, converted_attributes:)
+        errors = {}
+
+        invalid_attributes = activity.programme? ?
+          INVALID_LEVEL_B_ISPF_NON_ODA_ATTRIBUTES : INVALID_LEVEL_C_D_ISPF_NON_ODA_ATTRIBUTES
+
+        invalid_attributes.each do |invalid_attribute|
+          value = converted_attributes[invalid_attribute]
+
+          errors[invalid_attribute] = [value, I18n.t("importer.errors.activity.oda_attribute_in_non_oda_activity")] if value.present?
+        end
+
+        errors
+      end
+
       def is_oda_by_type(type:)
         {
           ispf_oda: true,
@@ -131,7 +146,18 @@ class Activity
       def update
         return unless @activity && @errors.empty?
 
-        @activity.assign_attributes(@converter.to_h)
+        attributes = @converter.to_h
+
+        if @is_oda == false
+          @errors.merge!(Activity::Import.invalid_non_oda_attribute_errors(
+            activity: @activity,
+            converted_attributes: attributes
+          ))
+        end
+
+        return unless @errors.blank?
+
+        @activity.assign_attributes(attributes)
 
         if @activity.sdg_1 || @activity.sdg_2 || @activity.sdg_3
           @activity.sdgs_apply = true
@@ -219,7 +245,19 @@ class Activity
         ) { |a|
           a.form_state = "complete"
         }
-        @activity.assign_attributes(@converter.to_h)
+
+        attributes = @converter.to_h
+
+        if @is_oda == false
+          @errors.merge!(Activity::Import.invalid_non_oda_attribute_errors(
+            activity: @activity,
+            converted_attributes: attributes
+          ))
+        end
+
+        return unless @errors.blank?
+
+        @activity.assign_attributes(attributes)
 
         if @activity.sdg_1 || @activity.sdg_2 || @activity.sdg_3
           @activity.sdgs_apply = true
@@ -585,6 +623,17 @@ class Activity
         raise I18n.t("importer.errors.activity.linked_activity_not_found") if linked_activity.nil?
 
         linked_activity.id
+      end
+
+      def convert_tags(tags)
+        return [] if tags.blank?
+
+        valid_codes = tags_options.map { |tag| tag.code.to_s }
+        tags.split("|").map do |tag|
+          raise I18n.t("importer.errors.activity.invalid_tags", code: tag) unless valid_codes.include?(tag)
+
+          Integer(tag)
+        end
       end
 
       def parse_date(date, message)